Conflicts:
	AUTHORS
	configure.ac
	db/zm_create.sql.in
	scripts/ZoneMinder/lib/ZoneMinder/ConfigData.pm.in
	scripts/zmupdate.pl.in
	src/Makefile.am
	src/zm_buffer.cpp
	src/zm_config_defines.h
	src/zm_event.cpp
	src/zm_ffmpeg.h
	src/zm_ffmpeg_camera.cpp
	src/zm_image.cpp
	src/zm_jpeg.h
	src/zm_local_camera.cpp
	src/zm_logger.cpp
	src/zm_mem_utils.h
	src/zm_monitor.cpp
	src/zm_monitor.h
	src/zm_mpeg.cpp
	src/zm_remote_camera_http.cpp
	src/zm_remote_camera_rtsp.cpp
	src/zm_rtsp.cpp
	src/zm_sdp.cpp
	src/zm_sdp.h
	src/zm_stream.cpp
	src/zm_thread.cpp
	src/zm_thread.h
	src/zm_threaddata.cpp
	src/zm_user.cpp
	src/zm_user.h
	src/zm_utils.cpp
	src/zm_zone.cpp
	src/zm_zone.h
	web/includes/actions.php
	web/lang/big5_big5.php
	web/lang/cn_zh.php
	web/lang/cs_cz.php
	web/lang/de_de.php
	web/lang/dk_dk.php
	web/lang/en_gb.php
	web/lang/es_ar.php
	web/lang/et_ee.php
	web/lang/fr_fr.php
	web/lang/he_il.php
	web/lang/hu_hu.php
	web/lang/it_it.php
	web/lang/ja_jp.php
	web/lang/nl_nl.php
	web/lang/pl_pl.php
	web/lang/pt_br.php
	web/lang/ro_ro.php
	web/lang/ru_ru.php
	web/lang/se_se.php
	web/skins/classic/js/skin.js
	web/skins/classic/views/js/monitor.js.php
	web/skins/classic/views/monitor.php
	web/skins/classic/views/zone.php

+++ src/zm_event.cpp
@@ -107,7 +107,7 @@ Event::Event( Monitor *p_monitor, struct timeval p_start_time, const std::string
         char date_path[PATH_MAX] = "";
         char time_path[PATH_MAX] = "";
         char *time_path_ptr = time_path;
-        for ( int i = 0; i < sizeof(dt_parts)/sizeof(*dt_parts); i++ )
+        for ( unsigned int i = 0; i < sizeof(dt_parts)/sizeof(*dt_parts); i++ )
         {
             path_ptr += snprintf( path_ptr, sizeof(path)-(path_ptr-path), "/%02d", dt_parts[i] );
 
@@ -589,7 +589,7 @@ void Event::AddFrame( Image *image, struct timeval timestamp, int score, Image *
         alarm_frames++;
 
         tot_score += score;
-        if ( score > max_score )
+        if ( score > (int)max_score )
             max_score = score;
 
         if ( alarm_image )
@@ -683,7 +683,7 @@ bool EventStream::loadInitialEventData( int monitor_id, time_t event_time )
         curr_frame_id = 1;
         if ( event_time >= event_data->start_time )
         {
-            for ( int i = 0; i < event_data->frame_count; i++ )
+            for (unsigned int i = 0; i < event_data->frame_count; i++ )
             {
                 //Info( "eft %d > et %d", event_data->frames[i].timestamp, event_time );
                 if ( event_data->frames[i].timestamp >= event_time )
@@ -1122,7 +1122,7 @@ void EventStream::checkEventLoaded()
         snprintf( sql, sizeof(sql), "select Id from Events where MonitorId = %ld and Id < %ld order by Id desc limit 1", event_data->monitor_id, event_data->event_id );
         reload_event = true;
     }
-    else if ( curr_frame_id > event_data->frame_count )
+    else if ( (unsigned int)curr_frame_id > event_data->frame_count )
     {
         snprintf( sql, sizeof(sql), "select Id from Events where MonitorId = %ld and Id > %ld order by Id asc limit 1", event_data->monitor_id, event_data->event_id );
         reload_event = true;
